# generated/swagger-ui/swagger.json
{
  "swagger" : "2.0",
  "info" : {
    "description" : "Working sample of Spring MVC annotations",
    "version" : "v1",
    "title" : "Swagger Maven Plugin Spring MVC working example",
    "termsOfService" : "http://www.github.com",
    "contact" : {
      "name" : "Name Surname",
      "url" : "http",
      "email" : "your@email.com"
    },
    "license" : {
      "name" : "License name",
      "url" : "http://www.license.com"
    }
  },
  "host" : "sample.com",
  "basePath" : "/api",
  "schemes" : [ "http", "https" ],
  "paths" : {
    "/app/rest/password/{password}" : {
      "post" : {
        "summary" : "Resets password",
        "description" : "",
        "operationId" : "resetPassword",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/app/users" : {
      "get" : {
        "summary" : "Retrieves users",
        "description" : "",
        "operationId" : "getUsers",
        "produces" : [ "application/json" ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    },
    "/app/users/{userId}/{someToken}" : {
      "post" : {
        "summary" : "Update user",
        "description" : "",
        "operationId" : "updateUser",
        "produces" : [ "application/json" ],
        "parameters" : [ {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "string"
        }, {
          "name" : "",
          "in" : "path",
          "required" : true,
          "type" : "string"
        } ],
        "responses" : {
          "200" : {
            "description" : "successful operation",
            "schema" : {
              "type" : "string"
            }
          }
        }
      }
    }
  }
}